Restructure the test suite.
Every concrete test in test_version.py and test_compress.py used to be
a thin wrapper around a “helper” method in tests/base.py, which made
extra work for adding new tests.  The setUp() logic was also split
between the tests_* files and base.py, making it hard to see what it
did or how it interacted with setup.py.  Finally, the 2.6
compatibility logic was clunkier than it needed to be.

Move all of the setUp() logic into base.py; conversely, move all of
the concrete tests into test_version.py and test_compress.py.
Reimplement test skipping in 2.6 as a backport of a subset of the 2.7
API, so none of the actual tests will have to change when 2.6 support
is dropped.  Improve readability of the very long strings used for
compression testing.

Exactly the same set of things is tested afterward as beforehand.