Add an interop test for the current compression format.
This verifies that the compression of a string at levels -5, 3, and 22
by the current compress() is byte-for-byte identical with the
compressed file produced by `zstd --no-check` at those levels.

Also reencode CDATA_036 and CDATA_046 with every byte as an \x escape,
which actually makes them easier to read, and also makes it less
likely that a search-and-replace over the whole file will corrupt them
by accident.